+++ src/main/java/com/falsepattern/zigbrains/lsp/client/languageserver/wrapper/LanguageServerWrapper.java
@@ -69,6 +69,7 @@ import org.eclipse.lsp4j.HoverCapabilities;
 import org.eclipse.lsp4j.InitializeParams;
 import org.eclipse.lsp4j.InitializeResult;
 import org.eclipse.lsp4j.InitializedParams;
+import org.eclipse.lsp4j.InlayHintCapabilities;
 import org.eclipse.lsp4j.OnTypeFormattingCapabilities;
 import org.eclipse.lsp4j.RangeFormattingCapabilities;
 import org.eclipse.lsp4j.ReferencesCapabilities;
@@ -578,6 +579,7 @@ public class LanguageServerWrapper {
         textDocumentClientCapabilities.setDocumentHighlight(new DocumentHighlightCapabilities());
         textDocumentClientCapabilities.setFormatting(new FormattingCapabilities());
         textDocumentClientCapabilities.setHover(new HoverCapabilities());
+        textDocumentClientCapabilities.setInlayHint(new InlayHintCapabilities());
         textDocumentClientCapabilities.setOnTypeFormatting(new OnTypeFormattingCapabilities());
         textDocumentClientCapabilities.setRangeFormatting(new RangeFormattingCapabilities());
         textDocumentClientCapabilities.setReferences(new ReferencesCapabilities());